Class-level tests
